backend/requirements.txt
crewai==0.130.0
crewai-tools>=0.9.0
litellm>=1.39.0
fastmcp>=2.0.0
uvicorn>=0.34.0
python-dotenv>=1.0.1
